chore(companion): review round-3 followups
- FocusModeButton: drop its duplicate fullscreenchange listener.
  Companion.tsx already owns the focus state and registers the
  same handler; one listener is enough.
- SetupMenu: tag input switches to defaultValue + key={session.id}
  so restoreFromArchive and newSession sync the textbox without
  needing a local useState.
- FocusExitButton: simplify props to onExit + onShowChrome. The
  chrome-visible branch was dead code (only mounted when
  chromeInFocus is false). Hide-peek now lives only in the bar.
- Extract DiceMenu out of CompanionBar so the bar drops back to
  185 lines, comfortably under the ~200-line guideline.
