+++ src/components/companion/FocusExitButton.tsx
@@ -1,12 +1,11 @@
 import { useEffect, useState } from "react";
-import { Eye, EyeOff, Minimize2 } from "lucide-react";
+import { Eye, Minimize2 } from "lucide-react";
 import { Button } from "@/components/ui/button";
 import { cn } from "@/lib/utils";
 
 interface FocusExitButtonProps {
   onExit: () => void;
-  chromeVisible: boolean;
-  onToggleChrome: (next: boolean) => void;
+  onShowChrome: () => void;
 }
 
 const AUTO_DIM_AFTER_MS = 4000;
@@ -16,9 +15,11 @@ const AUTO_DIM_AFTER_MS = 4000;
  * top-right inside the safe-area-inset; full opacity for the first few
  * seconds, then dims to a small always-tappable chip. Houses two
  * actions: peek the bar + phase strip back in (without leaving focus
- * mode) and exit focus entirely.
+ * mode) and exit focus entirely. Only mounted while focus is on AND
+ * the chrome is hidden — the hide-peek control lives in the bar
+ * itself once chrome is visible.
  */
-export function FocusExitButton({ onExit, chromeVisible, onToggleChrome }: FocusExitButtonProps) {
+export function FocusExitButton({ onExit, onShowChrome }: FocusExitButtonProps) {
   const [bright, setBright] = useState(true);
 
   useEffect(() => {
@@ -50,12 +51,12 @@ export function FocusExitButton({ onExit, chromeVisible, onToggleChrome }: Focus
       <Button
         size="sm"
         variant="ghost"
-        onClick={() => onToggleChrome(!chromeVisible)}
+        onClick={onShowChrome}
         className="h-9 gap-1 rounded-full px-3 text-xs font-semibold"
-        aria-label={chromeVisible ? "Hide controls" : "Show controls"}
-        title={chromeVisible ? "Hide bar and phase strip" : "Peek bar and phase strip"}
+        aria-label="Show controls"
+        title="Peek bar and phase strip"
       >
-        {chromeVisible ? <EyeOff className="size-4" /> : <Eye className="size-4" />}
+        <Eye className="size-4" />
       </Button>
       <Button
         size="sm"
